Revert "fix: moved sagas directory to store"
