+++ src/sagas/requests.js
@@ -1,6 +1,6 @@
 import { takeEvery, call, put } from 'redux-saga/effects';
 import Axios from 'axios';
-import { GET_REQUESTS, GET_REQUESTS_SUCCESS, GET_REQUESTS_FAILURE } from '../actions/types';
+import { GET_REQUESTS, GET_REQUESTS_SUCCESS, GET_REQUESTS_FAILURE } from '../store/actions/types';
 
 function* workerGetCrossCheckSessions() {
   const uri = 'http://localhost:3000/requests';
