+++ src/store/store.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import logger from 'redux-logger';
 import reduxSaga from 'redux-saga';
-import rootSaga from './sagas';
+import rootSaga from '../sagas';
 import rootReducer from './reducers';
 
 const reduxSagaMiddleware = reduxSaga();
